server/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<4.0
django-cors-headers>=4.6,<5.0
channels>=4.2,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0
psycopg2-binary>=2.9,<3.0
